+++ packages/terra-legacy-theme/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+2.18.0 - (July 19, 2018)
+------------------
+### Changed
+* Add .module extension to test SCSS file(s)
+
+2.17.0 - (July 17, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.16.0 - (July 10, 2018)
 ------------------
 ### Changed